+++ cohere/client_async.py
@@ -4,6 +4,7 @@ import os
 import posixpath
 import time
 from collections import defaultdict
+from dataclasses import asdict
 from functools import partial
 from typing import Any, Callable, Dict, List, Optional, Union
 
@@ -23,9 +24,11 @@ from cohere.responses import (
     Detokenization,
     Embeddings,
     GenerateFeedbackResponse,
+    GeneratePreferenceFeedbackResponse,
     Generations,
     LabelPrediction,
     Language,
+    PreferenceRating,
     Reranking,
     StreamingGenerations,
     SummarizeResponse,
@@ -60,7 +63,7 @@ class AsyncClient(Client):
         timeout=120,
     ) -> None:
         self.api_key = api_key or os.getenv("CO_API_KEY")
-        self.api_url = cohere.COHERE_API_URL
+        self.api_url = os.getenv("CO_API_URL", cohere.COHERE_API_URL)
         self.batch_size = cohere.COHERE_EMBED_BATCH_SIZE
         self.num_workers = num_workers
         self.request_dict = request_dict
@@ -340,20 +343,41 @@ class AsyncClient(Client):
         response = await self._request(cohere.GENERATE_FEEDBACK_URL, json_body)
         return GenerateFeedbackResponse(id=response["id"])
 
+    async def generate_preference_feedback(
+        self,
+        ratings: List[PreferenceRating],
+        model=None,
+        prompt: str = None,
+        annotator_id: str = None,
+    ) -> GeneratePreferenceFeedbackResponse:
+        ratings_dicts = []
+        for rating in ratings:
+            ratings_dicts.append(asdict(rating))
+        json_body = {
+            "ratings": ratings_dicts,
+            "prompt": prompt,
+            "annotator_id": annotator_id,
+            "model": model,
+        }
+        response = await self._request(cohere.GENERATE_PREFERENCE_FEEDBACK_URL, json_body)
+        return GeneratePreferenceFeedbackResponse(id=response["id"])
+
     async def rerank(
         self,
         query: str,
         documents: Union[List[str], List[Dict[str, Any]]],
-        model: str = None,
+        model: str,
         top_n: Optional[int] = None,
+        max_chunks_per_doc: Optional[int] = None,
     ) -> Reranking:
         """Returns an ordered list of documents ordered by their relevance to the provided query
 
         Args:
             query (str): The search query
             documents (list[str], list[dict]): The documents to rerank
-            model (str): (Optional) The model to use for re-ranking
+            model (str): The model to use for re-ranking
             top_n (int): (optional) The number of results to return, defaults to returning all results
+            max_chunks_per_doc (int): (optional) The maximum number of chunks derived from a document
         """
         parsed_docs = []
         for doc in documents:
@@ -372,6 +396,7 @@ class AsyncClient(Client):
             "model": model,
             "top_n": top_n,
             "return_documents": False,
+            "max_chunks_per_doc": max_chunks_per_doc,
         }
         reranking = Reranking(await self._request(cohere.RERANK_URL, json=json_body))
         for rank in reranking.results:
@@ -391,10 +416,10 @@ class AsyncClient(Client):
             embeddings_url (str): File with embeddings to cluster.
             min_cluster_size (Optional[int], optional): Minimum number of elements in a cluster. Defaults to 10.
             n_neighbors (Optional[int], optional): Number of nearest neighbors used by UMAP to establish the
-            local structure of the data. Defaults to 15. For more information, please refer to
-            https://umap-learn.readthedocs.io/en/latest/parameters.html#n-neighbors
+                local structure of the data. Defaults to 15. For more information, please refer to
+                https://umap-learn.readthedocs.io/en/latest/parameters.html#n-neighbors
             is_deterministic (Optional[bool], optional): Determines whether the output of the cluster job is
-            deterministic. Defaults to True.
+                deterministic. Defaults to True.
 
         Returns:
             CreateClusterJobResponse: Created clustering job handler