+++ tests/async/test_async_cluster.py
@@ -90,6 +90,7 @@ async def test_async_job_wait_method_succeeds(async_client: AsyncClient):
 
 
 @pytest.mark.asyncio
+@pytest.mark.skipif(IN_CI, reason="can timeout during high load")
 async def test_async_job_wait_method_times_out(async_client: AsyncClient):
     create_res = await async_client.create_cluster_job(
         INPUT_FILE,