+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 # Changelog
 
+## 4.1.6
+
+- [#205](https://github.com/cohere-ai/cohere-python/pull/205)
+    - Add param max_chunks_per_doc to rerank
+    - Enforce model param for rerank
+
+## 4.1.5
+
+- [#208](https://github.com/cohere-ai/cohere-python/pull/208)
+  - Fix a missing import for CohereConnectionError
+
+## 4.1.4
+
+- [#204](https://github.com/cohere-ai/cohere-python/pull/204)
+  - Add `generate_preference_feedback` for submitting preference-style feedback
+
 ## 4.1.3
+
 - [#194](https://github.com/cohere-ai/cohere-python/pull/194)
-   - Return the generation ID for chat
+  - Return the generation ID for chat
 
 ## 4.1.2